backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pgvector==0.2.4
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-docx==1.1.0
pypdf==4.0.1
openpyxl==3.1.2
python-pptx==0.6.23
markdown==3.5.2
minio==7.2.3
docker==7.0.0
httpx==0.26.0
aiofiles==23.2.1
jinja2==3.1.3
weasyprint==60.2
pandas==2.2.0
numpy==1.26.3
matplotlib==3.8.2
plotly==5.18.0
